+++ NEWS.md
@@ -1,3 +1,9 @@
+# BioDataScience 2021.2.0
+
+- `render_sdd()` function convert a Rmarkdown files to html document with specific arguments.
+- `corr_rmd()` function is used to try to convert all Rmarkdown files in a project to html documents. If a document does not compile, a copy is created.
+- `Check my project` addin is an Rstudio addin used to validate several criteria on a project like the compilation of Rmarkdown.
+
 # BioDataScience 2021.1.0
 
 - `sdd_info()` function and corresponfing RStudio addin are introduced for easily sharing small pieces of text in a central repository.